rebuild/requirements.txt
Flask==3.1.1
huggingface-hub==0.33.1
numpy==2.2.6
opencv-python-headless==4.11.0.86
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
supervision==0.25.1
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
ultralytics==8.3.161
ultralytics-thop==2.0.14